move the export filename input to a different widget
